+++ dbms/src/Columns/ColumnUtils.h
@@ -15,8 +15,10 @@
 #pragma once
 
 #include <Columns/IColumn.h>
+#include <Core/ColumnWithTypeAndName.h>
 
 namespace DB
 {
 bool columnEqual(const ColumnPtr & expected, const ColumnPtr & actual, String & unequal_msg);
+void convertColumnToNullable(ColumnWithTypeAndName & column);
 } // namespace DB
